pyproject.toml
[project]
name = "blackjax-lumen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anesthetic>=2.13.1",
    "blackjax @ git+https://github.com/handley-lab/blackjax@nested_sampling",
    "distrax",
    "h5py>=3.16.0",
    "jax-bandflux>=1.0.0",
    "jupyter>=1.1.1",
    "lab>=8.9",
    "matplotlib>=3.10.8",
    "pandas<3",
    "tqdm>=4.67.3",
]

[tool.uv.sources]
distrax = { git = "https://github.com/google-deepmind/distrax" }
